Update swagger dir and name
