+++ src/utils/index.ts
@@ -1,0 +1,3 @@
+import swaggerSpec from "./swagger.util";
+
+export { swaggerSpec };
